CHROMIUM: driver: media: virtio: Don't query bitrate for decoder.
Upon opening the virtio device we currently query for
VIRTIO_VIDEO_CONTROL_BITRATE for both decoder and encoder devices. For
a decoder querying this value doesn't make any sense, so this CL makes
sure we only query the bitrate when the device is an encoder.

Signed-off-by: David Staessens <dstaessens@google.com>

BUG=b:175512726
TEST=tast run DUT arc.arc.VideoDecodeAccel.h264_vm

Change-Id: I198ebff874429a8b2e36b2703de0cba8aa398d40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2944505
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>